Draw the sky into the pack's framebuffer, and strip the VTune profiling
The sky stage ran between renderBegin() and bindScene(), so gbuffers_skybasic
and gbuffers_skytextured wrote into whatever framebuffer the begin passes left
bound and were thrown away. Every external shader pack lost its sun, moon and
stars; vanilla, which has no scene framebuffer to miss, was the one thing that
looked right. Iris rebinds the program's own framebuffer on every gbuffer
program bind (ExtendedShader.apply -> writingToBeforeTranslucent.bind()), and
bindScene() is the one place that does it here. Move the sky stage after the
shadowcomp / bindScene / prepare group, which is also Iris' order:
renderShadows() ends with prepareRenderer.renderAll(), and
WorldRenderingPhase.SKY follows it.

Fog class now comes from the kind of draw rather than the resolved program name,
matching Iris' ShaderKey table. SKY_BASIC keeps FogMode.OFF even when
gbuffers_skybasic falls back through gbuffers_basic, so a sky program no longer
inherits fog from whatever it resolved to; gbuffers_textured and
gbuffers_damagedblock are OFF with it, and the shadow pass is OFF for every id.

Profiling and debug scaffolding is out ahead of publishing: the VTune ITT header
and every VT_TRACE call site, the VTUNE_ENABLED CMake option and its per-target
wiring, the -Vtune build switch, scripts/vtune-profile.ps1, and
FramePipeline::phaseName(), which only ever produced trace labels. The F3
ClientProfilerOverlay is a shipped game feature and stays.

This also publishes the work accumulated in the working tree since the last
commit and never pushed: the mina audio backend, terrain capture/upload and
culling, chunk and lighting changes, shader pack loading and uniforms, input,
GUI and screen changes, world and chunk storage, the Lua mod runtime, six new
test files, and the spc and worldedit mods. item_drop_physics is removed.

+++ src/net/minecraft/client/input/InputSystem.hpp
@@ -60,10 +60,10 @@ class InputSystem {
  void lockCursor();
  void unlockCursor();
  void pollMouseLook();
- [[nodiscard]] int mouseLookDeltaX() const noexcept {
+ [[nodiscard]] double mouseLookDeltaX() const noexcept {
   return mouseLookDeltaX_;
  }
- [[nodiscard]] int mouseLookDeltaY() const noexcept {
+ [[nodiscard]] double mouseLookDeltaY() const noexcept {
   return mouseLookDeltaY_;
  }
  [[nodiscard]] bool isMouseButtonDown(int button) const;
@@ -91,10 +91,10 @@ class InputSystem {
  std::array<bool, 3> mouseDown_{};
  int cursorX_ = 0;
  int cursorY_ = 0;
- int mouseLastX_ = 0;
- int mouseLastY_ = 0;
+ double mouseLastX_ = 0.0;
+ double mouseLastY_ = 0.0;
  bool mouseHasLastPoint_ = false;
- int mouseLookDeltaX_ = 0;
- int mouseLookDeltaY_ = 0;
+ double mouseLookDeltaX_ = 0.0;
+ double mouseLookDeltaY_ = 0.0;
 };
 } // namespace net::minecraft::client::input
